move boto out of staging credentials
